ONEAPI-529: compute the page offset as a long, not an int
pageNo validates as 1..Integer.MAX_VALUE, so pageNo=5,000,000 at pageSize=500
gives an offset of 2,499,999,500 — which wraps to -1,794,967,796 as an int and
puts a negative OFFSET into the statement. That is a request error rather than
the empty page it should be.

PageRequest.getOffset() returned a long and was absorbing this; dropping the
Pageable in slice 2 dropped that with it. Found by self-review, not by a test —
the differential drives the statement directly and never computes an offset
from a page number.

+++ src/main/java/com/nextgen/gameaggregator/repository/ga/reader/VendorGameReaderRepository.java
@@ -133,7 +133,7 @@ public interface VendorGameReaderRepository extends JpaRepository<VendorGame, In
             @Param("masterAgentId") Integer masterAgentId,
             @Param("agentId") Integer agentId,
             @Param("limit") Integer limit,
-            @Param("offset") Integer offset);
+            @Param("offset") Long offset);
 
     // ONEAPI-529: counts games over GAME_LIST_FILTER, the same filter the list pages, so the
     // total and the pages cannot disagree. No join multiplication and no GROUP BY -- the
